refactor(test): extract helper functions
