+++ config/plugin-commands-config/test/configSet.test.ts
@@ -4,70 +4,8 @@ import { PnpmError } from '@pnpm/error'
 import { tempDir } from '@pnpm/prepare'
 import { config } from '@pnpm/plugin-commands-config'
 import { readIniFileSync } from 'read-ini-file'
-import { writeIniFileSync } from 'write-ini-file'
 import { sync as readYamlFile } from 'read-yaml-file'
-import { sync as writeYamlFile } from 'write-yaml-file'
-
-interface ConfigFilesData {
-  globalRc: Record<string, unknown> | undefined
-  globalYaml: Record<string, unknown> | undefined
-  localRc: Record<string, unknown> | undefined
-  localYaml: Record<string, unknown> | undefined
-}
-
-function readConfigFiles (globalConfigDir: string | undefined, localDir: string | undefined): ConfigFilesData {
-  function tryRead<Return> (reader: () => Return): Return | undefined {
-    try {
-      return reader()
-    } catch (error) {
-      if (error && typeof error === 'object' && 'code' in error && error.code === 'ENOENT') {
-        return undefined
-      }
-      throw error
-    }
-  }
-
-  return {
-    globalRc: globalConfigDir
-      ? tryRead(() => readIniFileSync(path.join(globalConfigDir, 'rc')) as Record<string, unknown>)
-      : undefined,
-    globalYaml: globalConfigDir
-      ? tryRead(() => readYamlFile(path.join(globalConfigDir, 'rc.yaml')))
-      : undefined,
-    localRc: localDir
-      ? tryRead(() => readIniFileSync(path.join(localDir, '.npmrc')) as Record<string, unknown>)
-      : undefined,
-    localYaml: localDir
-      ? tryRead(() => readYamlFile(path.join(localDir, 'pnpm-workspace.yaml')))
-      : undefined,
-  }
-}
-
-function writeConfigFiles (globalConfigDir: string | undefined, localDir: string | undefined, data: ConfigFilesData): void {
-  if (globalConfigDir) {
-    fs.mkdirSync(globalConfigDir, { recursive: true })
-
-    if (data.globalRc) {
-      writeIniFileSync(path.join(globalConfigDir, 'rc'), data.globalRc)
-    }
-
-    if (data.globalYaml) {
-      writeYamlFile(path.join(globalConfigDir, 'rc.yaml'), data.globalYaml)
-    }
-  }
-
-  if (localDir) {
-    fs.mkdirSync(localDir, { recursive: true })
-
-    if (data.localRc) {
-      writeIniFileSync(path.join(localDir, '.npmrc'), data.localRc)
-    }
-
-    if (data.localYaml) {
-      writeYamlFile(path.join(localDir, 'pnpm-workspace.yaml'), data.localYaml)
-    }
-  }
-}
+import { type ConfigFilesData, readConfigFiles, writeConfigFiles } from './utils/index.js'
 
 test('config set registry setting using the global option', async () => {
   const tmp = tempDir()
